+++ internal/sdkv2provider/resource_cloudflare_zone_hold_test.go
@@ -38,9 +38,10 @@ func TestAccCloudflareZoneHold_Full(t *testing.T) {
 				),
 			},
 			{
-				ResourceName:      name,
-				ImportState:       true,
-				ImportStateVerify: true,
+				ResourceName:            name,
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"hold_after"},
 			},
 		},
 	})